+++ packages/directives/vLoading.ts
@@ -21,13 +21,13 @@ export default {
         if (el.originalPosition !== "absolute" && el.originalPosition !== "fixed") {
           el.style.position = "relative";
         }
-        el.comp.show = true;
+        el.comp.cShow = true;
 
         if(title){
-          el.comp.title = title;
+          el.comp.cTitle = title;
         }
         if(size){
-          el.comp.size = size;
+          el.comp.cSize = size;
         }
         el.comp.root= (app._instance as ComponentInternalInstance)?.root?.proxy;
         el?.appendChild(el?.loadingRoot);
@@ -63,7 +63,7 @@ export default {
         });
         el.domInserted = true;
       } else {
-        el.comp.show = false;
+        el.comp.cShow = false;
         if (el.domInserted) {
           removeDom(el);
         }