+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 本文件格式基于 [Keep a Changelog](https://keepachangelog.com/zh-CN/1.0.0/) ，并且遵循 [Semantic Versioning](https://semver.org/spec/v2.0.0.html) 。
 
+
+## [0.0.19] - 2021-08-17
+
+### Added
+
+- EASIButton 新增info、success、danger、warning四种颜色状态
+
+### Changed
+
+- EASILoading Props改为show(v-model)、size、title
+
 ## [0.0.18] - 2021-08-17
 
 ### Added